# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/books": {
            "get": {
                "description": "Возвращает список всех книг в хранилище",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Получить все книги",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет описание книги и сохраняет в хранилище",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Обновить книгу",
                "parameters": [
                    {
                        "description": "Данные для обновления книги",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт объект описания книги и сохраняет в хранилище",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Создать книгу",
                "parameters": [
                    {
                        "description": "Данные для создания книги",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/books/count": {
            "put": {
                "description": "Меняет количество копий конкретной книги в хранилище",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Изменить количество",
                "parameters": [
                    {
                        "description": "Id книги и количество для изменения (может быть больше или меньше нуля)",
                        "name": "quantity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeBookQuantityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ChangeBookQuantityResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/books/search": {
            "get": {
                "description": "Ищет книгу по переданным в query параметрам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Поиск книги по параметрам",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название книги",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя автора (ФИО или Фамилия)",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Год написания книги (формат YYYY-MM-DD)",
                        "name": "yearOfWriting",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Год рождения автора (формат YYYY-MM-DD)",
                        "name": "yearOfBirth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/books/{book_id}": {
            "get": {
                "description": "Ищет книгу по id в хранилище",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Поиск книги по id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{login}": {
            "get": {
                "description": "Возвращает информацию о пользователе по его логину",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Найти пользователя по логину",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя (UUID)",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Author": {
            "type": "object",
            "required": [
                "author_id",
                "birth",
                "first_name",
                "second_name",
                "surname"
            ],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "second_name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "models.Book": {
            "type": "object",
            "required": [
                "author",
                "book_id",
                "quantity",
                "title",
                "year"
            ],
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.Author"
                },
                "book_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "models.ChangeBookQuantityRequest": {
            "type": "object",
            "required": [
                "bookId",
                "quantity"
            ],
            "properties": {
                "bookId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.ChangeBookQuantityResponse": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "models.CreateBookRequest": {
            "type": "object",
            "required": [
                "author",
                "title",
                "year"
            ],
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.Author"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "models.CreateBookResponse": {
            "type": "object",
            "required": [
                "bookId"
            ],
            "properties": {
                "bookId": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateBookRequest": {
            "type": "object",
            "required": [
                "author",
                "book_id",
                "title",
                "year"
            ],
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.Author"
                },
                "book_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}